+++ Route4MeSDKLibrary/DataTypes/AddressManifest.cs
@@ -12,19 +12,19 @@ namespace Route4MeSDK.DataTypes
         /// How much time is to be spent on service from the start in seconds.
         /// </summary>
         [DataMember(Name = "running_service_time", EmitDefaultValue = false)]
-        public int? RunningServiceTime { get; set; }
+        public long? RunningServiceTime { get; set; }
 
         /// <summary>
         /// How much time is spent driving from the start in seconds.
         /// </summary>
         [DataMember(Name = "running_travel_time", EmitDefaultValue = false)]
-        public int? RunningTravelTime { get; set; }
+        public long? RunningTravelTime { get; set; }
 
         /// <summary>
         /// Running wait time.
         /// </summary>
         [DataMember(Name = "running_wait_time", EmitDefaultValue = false)]
-        public int? RunningWaitTime { get; set; }
+        public long? RunningWaitTime { get; set; }
 
         /// <summary>
         /// Distance traversed before reaching this address.
@@ -48,52 +48,52 @@ namespace Route4MeSDK.DataTypes
         /// Projected arrival time UTC unixtime.
         /// </summary>
         [DataMember(Name = "projected_arrival_time_ts", EmitDefaultValue = false)]
-        public int? ProjectedArrivalTimeTs { get; set; }
+        public long? ProjectedArrivalTimeTs { get; set; }
 
         /// <summary>
         /// Estimated departure time UTC unixtime.
         /// </summary>
         [DataMember(Name = "projected_departure_time_ts", EmitDefaultValue = false)]
-        public int? ProjectedDepartureTimeTs { get; set; }
+        public long? ProjectedDepartureTimeTs { get; set; }
 
         /// <summary>
         /// Time when the address was marked as visited UTC unixtime. 
         /// This is actually equal to timestamp_last_visited most of the time.
         /// </summary>
         [DataMember(Name = "actual_arrival_time_ts", EmitDefaultValue = false)]
-        public int? ActualArrivalTimeTs { get; set; }
+        public long? ActualArrivalTimeTs { get; set; }
 
         /// <summary>
         /// Time when the address was mared as departed UTC. 
         /// This is actually equal to timestamp_last_departed most of the time.
         /// </summary>
         [DataMember(Name = "actual_departure_time_ts", EmitDefaultValue = false)]
-        public int? ActualDepartureTimeTs { get; set; }
+        public long? ActualDepartureTimeTs { get; set; }
 
         /// <summary>
         /// Estimated arrival time based on the current route progress, 
         /// i.e. based on the last known actual_arrival_time.
         /// </summary>
         [DataMember(Name = "estimated_arrival_time_ts", EmitDefaultValue = false)]
-        public int? EstimatedArrivalTimeTs { get; set; }
+        public long? EstimatedArrivalTimeTs { get; set; }
 
         /// <summary>
         /// Estimated departure time based on the current route progress.
         /// </summary>
         [DataMember(Name = "estimated_departure_time_ts", EmitDefaultValue = false)]
-        public int? EstimatedDepartureTimeTs { get; set; }
+        public long? EstimatedDepartureTimeTs { get; set; }
 
         /// <summary>
         /// Scheduled arrival time. 
         /// </summary>
         [DataMember(Name = "scheduled_arrival_time_ts", EmitDefaultValue = false)]
-        public int? ScheduledArrivalTimeTs { get; set; }
+        public long? ScheduledArrivalTimeTs { get; set; }
 
         /// <summary>
         /// Scheduled departure time.
         /// </summary>
         [DataMember(Name = "scheduled_departure_time_ts", EmitDefaultValue = false)]
-        public int? ScheduledDepartureTimeTs { get; set; }
+        public long? ScheduledDepartureTimeTs { get; set; }
 
         /// <summary>
         /// This is the difference between the originally projected arrival time and Actual Arrival Time.
