+++ Route4MeSDKLibrary/DataTypes/FindAssetResponse.cs
@@ -133,13 +133,13 @@ namespace Route4MeSDK.DataTypes
         /// UNIX timestamp when a geofence visited event was triggered.
         /// </summary>
         [DataMember(Name = "timestamp_geofence_visited")]
-        public int? TimestampGeofenceVisited { get; set; }
+        public long? TimestampGeofenceVisited { get; set; }
 
         /// <summary>
         /// UNIX timestamp of a last visited event.
         /// </summary>
         [DataMember(Name = "timestamp_last_visited")]
-        public int? TimestampLastVisited { get; set; }
+        public long? TimestampLastVisited { get; set; }
     }
 
     /// <summary>
@@ -232,7 +232,7 @@ namespace Route4MeSDK.DataTypes
         /// Status getting timestamp
         /// </summary>
         [DataMember(Name = "unix_timestamp")]
-        public int? UnixTimestamp { get; set; }
+        public long? UnixTimestamp { get; set; }
 
         /// <summary>
         /// nformation about a shipped package.
